Refactor backend: add start script, implement test API route, and remove unused files
