+++ backend/src/supabase/supabaseClient.js
@@ -1,0 +1,15 @@
+// environment variables
+import "dotenv/config";
+
+import { createClient } from "@supabase/supabase-js";
+
+const supabaseUrl = process.env.SUPABASE_URL;
+const supabaseKey = process.env.SUPABASE_KEY;
+
+if (!supabaseUrl || !supabaseKey) {
+  throw new Error("Missing SUPABASE_URL or SUPABASE_KEY in .env file");
+}
+
+const supabase = createClient(supabaseUrl, supabaseKey);
+
+export default supabase;
